+++ torchtune/data/_utils.py
@@ -4,11 +4,16 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-import warnings
-from functools import lru_cache, wraps
-from typing import Any, Callable, Dict, List, Optional, TypeVar
+from typing import Any, Dict, List, Optional, TypeVar
+
+from torchtune.config._utils import _get_component_from_path
 
 from torchtune.data._messages import Message
+from torchtune.data._prompt_templates import (
+    _TemplateType,
+    PromptTemplate,
+    PromptTemplateInterface,
+)
 
 T = TypeVar("T", bound=type)
 
@@ -120,32 +125,30 @@ def validate_messages(
         last_turn = message.role
 
 
-def deprecated(msg: str = "") -> Callable[[T], T]:
+def _get_prompt_template(
+    prompt_template: _TemplateType,
+) -> PromptTemplateInterface:
     """
-    Decorator to mark an object as deprecated and print additional message.
+    Retrieve prompt template from import dotpath or create a custom one with provided
+    template dictionary.
 
     Args:
-        msg (str): additional information to print after warning.
+        prompt_template (_TemplateType): optional specified prompt template.
+            If a string, it is assumed to be the dotpath of a :class:`~torchtune.data.PromptTemplateInterface`
+            class. If a dictionary, it is assumed to be a custom prompt template mapping role to the
+            prepend/append tags.
 
     Returns:
-        Callable[[T], T]: the decorated object.
+        PromptTemplateInterface: the specified prompt template
+
+    Raises:
+        ValueError: If a string or dictionary is not passed in
     """
-
-    @lru_cache(maxsize=1)
-    def warn(obj):
-        warnings.warn(
-            f"{obj.__name__} is deprecated and will be removed in future versions. "
-            + msg,
-            category=FutureWarning,
-            stacklevel=3,
+    if isinstance(prompt_template, str):
+        return _get_component_from_path(prompt_template)()
+    elif isinstance(prompt_template, dict):
+        return PromptTemplate(prompt_template)
+    else:
+        raise ValueError(
+            f"Prompt template must be a dotpath string or dictionary with custom template, got {type(prompt_template)}"
         )
-
-    def decorator(obj):
-        @wraps(obj)
-        def wrapper(*args, **kwargs):
-            warn(obj)
-            return obj(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
