+++ torchtune/datasets/_instruct.py
@@ -16,10 +16,10 @@ from torchtune.data import (
     Message,
     validate_messages,
 )
-from torchtune.data._utils import deprecated
 from torchtune.datasets._packed import PackedDataset
 from torchtune.datasets._sft import SFTDataset
 from torchtune.modules.tokenizers import ModelTokenizer
+from torchtune.utils.logging import deprecated
 
 
 @deprecated(
