+++ torchtune/models/qwen2/_model_builders.py
@@ -11,7 +11,7 @@ from torchtune.modules import TransformerDecoder, TiedEmbeddingTransformerDecode
 from torchtune.modules.peft import LORA_ATTN_MODULES
 from torchtune.modules.tokenizers import parse_hf_tokenizer_json
 from torchtune.data._prompt_templates import _TemplateType
-from torchtune.config._utils import _get_prompt_template
+from torchtune.data._utils import _get_prompt_template
 
 """
 Model builders build specific instantiations using component builders. For example
